web/requirements.txt
pillow
docker
djangorestframework
django-filter
django-debug-toolbar